+++ application/core/Bancha_Output.php
@@ -23,6 +23,18 @@ class Bancha_Output extends CI_Output
 	}
 
 	/**
+	 * Build the file path.
+	 * The file name is an MD5 hash of the full URI plus the current theme.
+	 * @param string $uri
+	 * @param string $theme
+	 * @return string
+	 */
+	public function get_cachefile($uri, $theme)
+	{
+		return md5($uri) . '.' . $theme . '.tmp';
+	}
+
+	/**
 	 * Write a Cache File
 	 * We added the query string to the cache file name to include GET req and the .tmp extension
 	 * We also toggled off the caching system for stage contents
@@ -39,7 +51,7 @@ class Bancha_Output extends CI_Output
 			$CI->load->content();
 		}
 
-		if ($CI->content->is_stage || isset($_COOKIE['prevent_cache']))
+		if ($CI->content->is_stage || isset($_SERVER['prevent_cache']))
 		{
 			//Stage contents cannot be saved to disk because they will
 			//overwrite the production files
@@ -56,9 +68,8 @@ class Bancha_Output extends CI_Output
 			return;
 		}
 
-		$uri =	$_SERVER['REQUEST_URI'];
-
-		$cache_path .= md5($uri).'.tmp';
+		$cache_path .= $this->get_cachefile($_SERVER['REQUEST_URI'], $_SESSION['_website_theme']);
+		
 
 		if ( ! $fp = @fopen($cache_path, FOPEN_WRITE_CREATE_DESTRUCTIVE))
 		{
@@ -96,7 +107,8 @@ class Bancha_Output extends CI_Output
 	 */
 	public function _display_cache(&$CFG, &$URI)
 	{
-		if (isset($_COOKIE['prevent_cache']))
+
+		if (isset($_SERVER['prevent_cache']))
 		{
 			//The user is logged in. Let's always output not-cached pages
 			return FALSE;
@@ -104,10 +116,13 @@ class Bancha_Output extends CI_Output
 
 		$cache_path = ($CFG->item('cache_path') == '') ? APPPATH.'cache/' : $CFG->item('cache_path');
 
-		// Build the file path.  The file name is an MD5 hash of the full URI
-		$uri =	$_SERVER['REQUEST_URI'];
+		if (!isset($_SESSION['_website_theme']))
+		{
+			//We cannot trust users that doesn't have a theme
+			return FALSE;
+		}
 
-		$filepath = $cache_path.md5($uri).'.tmp';
+		$filepath = $cache_path . $this->get_cachefile($_SERVER['REQUEST_URI'], $_SESSION['_website_theme']);
 
 		if ( ! @file_exists($filepath))
 		{